requirements.txt
fastapi
uvicorn
streamlit
requests
pydantic
python-multipart
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
langchain
langchain-google-genai
langchain-community
pytz